inference_api/requirements.txt
onnxruntime-gpu==1.17.1
opencv-python-headless>=4.9.0
numpy>=1.26.0,<2.0.0
Pillow>=10.2.0
fastapi
uvicorn
python-multipart